fix: change int to int16

+++ cpu.cs
@@ -5,8 +5,8 @@ namespace STR
     class CPU
     {
         public static Int16 pc = 1; // program counter
-        public static int[] mem = new int[256]; // 256 bytes
-        public static int[] reg = new int[16]; // 16 registers
+        public static Int16[] mem = new Int16[256]; // 256 bytes
+        public static Int16[] reg = new Int16[16]; // 16 registers
         public static bool running = true;
         public static Int16 cycles = 0;
         public static Int16 currentInstruction = 0;
